requirements.txt
pytest
pytest-mock
pytest-cov
python-telegram-bot
APScheduler
requests